+++ packages/auth/__tests__/providers/cognito/rememberDevice.test.ts
@@ -7,9 +7,10 @@ import { Amplify, fetchAuthSession } from '@aws-amplify/core';
 import { AuthError } from '../../../src/errors/AuthError';
 import { rememberDevice } from '../../../src/providers/cognito';
 import { UpdateDeviceStatusException } from '../../../src/providers/cognito/types/errors';
-import { updateDeviceStatus } from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider';
 import { tokenOrchestrator } from '../../../src/providers/cognito/tokenProvider';
 import { DeviceMetadata } from '../../../src/providers/cognito/tokenProvider/types';
+import { createUpdateDeviceStatusClient } from '../../../src/foundation/factories/serviceClients/cognitoIdentityProvider';
+import { createCognitoUserPoolEndpointResolver } from '../../../src/providers/cognito/factories';
 
 import { getMockError, mockAccessToken } from './testUtils/data';
 import { setUpGetConfig } from './testUtils/setUpGetConfig';
@@ -23,8 +24,9 @@ jest.mock('@aws-amplify/core/internals/utils', () => ({
 	isBrowser: jest.fn(() => false),
 }));
 jest.mock(
-	'../../../src/providers/cognito/utils/clients/CognitoIdentityProvider',
+	'../../../src/foundation/factories/serviceClients/cognitoIdentityProvider',
 );
+jest.mock('../../../src/providers/cognito/factories');
 jest.mock('../../../src/providers/cognito/tokenProvider');
 
 describe('rememberDevice', () => {
@@ -35,7 +37,13 @@ describe('rememberDevice', () => {
 	};
 	// assert mocks
 	const mockFetchAuthSession = fetchAuthSession as jest.Mock;
-	const mockUpdateDeviceStatus = updateDeviceStatus as jest.Mock;
+	const mockUpdateDeviceStatus = jest.fn();
+	const mockCreateUpdateDeviceStatusClient = jest.mocked(
+		createUpdateDeviceStatusClient,
+	);
+	const mockCreateCognitoUserPoolEndpointResolver = jest.mocked(
+		createCognitoUserPoolEndpointResolver,
+	);
 	const mockGetDeviceMetadata =
 		tokenOrchestrator.getDeviceMetadata as jest.Mock;
 
@@ -49,12 +57,16 @@ describe('rememberDevice', () => {
 	beforeEach(() => {
 		mockGetDeviceMetadata.mockResolvedValue(mockDeviceMetadata);
 		mockUpdateDeviceStatus.mockResolvedValue({ $metadata: {} });
+		mockCreateUpdateDeviceStatusClient.mockReturnValueOnce(
+			mockUpdateDeviceStatus,
+		);
 	});
 
 	afterEach(() => {
 		mockGetDeviceMetadata.mockReset();
 		mockUpdateDeviceStatus.mockReset();
 		mockFetchAuthSession.mockClear();
+		mockCreateUpdateDeviceStatusClient.mockClear();
 	});
 
 	it('should call updateDeviceStatus client with correct request', async () => {
@@ -69,6 +81,25 @@ describe('rememberDevice', () => {
 			}),
 		);
 		expect(mockUpdateDeviceStatus).toHaveBeenCalledTimes(1);
+	});
+
+	it('invokes mockCreateCognitoUserPoolEndpointResolver with expected endpointOverride', async () => {
+		const expectedUserPoolEndpoint = 'https://my-custom-endpoint.com';
+		jest.mocked(Amplify.getConfig).mockReturnValueOnce({
+			Auth: {
+				Cognito: {
+					userPoolClientId: '111111-aaaaa-42d8-891d-ee81a1549398',
+					userPoolId: 'us-west-2_zzzzz',
+					identityPoolId: 'us-west-2:xxxxxx',
+					userPoolEndpoint: expectedUserPoolEndpoint,
+				},
+			},
+		});
+		await rememberDevice();
+
+		expect(mockCreateCognitoUserPoolEndpointResolver).toHaveBeenCalledWith({
+			endpointOverride: expectedUserPoolEndpoint,
+		});
 	});
 
 	it('should throw an error when service returns an error response', async () => {