+++ packages/auth/__tests__/providers/cognito/setUpTOTP.test.ts
@@ -6,8 +6,9 @@ import { decodeJWT } from '@aws-amplify/core/internals/utils';
 
 import { AuthError } from '../../../src/errors/AuthError';
 import { AssociateSoftwareTokenException } from '../../../src/providers/cognito/types/errors';
-import { associateSoftwareToken } from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider';
 import { setUpTOTP } from '../../../src/providers/cognito';
+import { createAssociateSoftwareTokenClient } from '../../../src/foundation/factories/serviceClients/cognitoIdentityProvider';
+import { createCognitoUserPoolEndpointResolver } from '../../../src/providers/cognito/factories';
 
 import { getMockError, mockAccessToken } from './testUtils/data';
 import { setUpGetConfig } from './testUtils/setUpGetConfig';
@@ -21,14 +22,21 @@ jest.mock('@aws-amplify/core/internals/utils', () => ({
 	isBrowser: jest.fn(() => false),
 }));
 jest.mock(
-	'../../../src/providers/cognito/utils/clients/CognitoIdentityProvider',
+	'../../../src/foundation/factories/serviceClients/cognitoIdentityProvider',
 );
+jest.mock('../../../src/providers/cognito/factories');
 
 describe('setUpTOTP', () => {
 	const secretCode = 'secret-code';
 	// assert mocks
 	const mockFetchAuthSession = fetchAuthSession as jest.Mock;
-	const mockAssociateSoftwareToken = associateSoftwareToken as jest.Mock;
+	const mockAssociateSoftwareToken = jest.fn();
+	const mockCreateAssociateSoftwareTokenClient = jest.mocked(
+		createAssociateSoftwareTokenClient,
+	);
+	const mockCreateCognitoUserPoolEndpointResolver = jest.mocked(
+		createCognitoUserPoolEndpointResolver,
+	);
 
 	beforeAll(() => {
 		setUpGetConfig(Amplify);
@@ -42,11 +50,15 @@ describe('setUpTOTP', () => {
 			SecretCode: secretCode,
 			$metadata: {},
 		});
+		mockCreateAssociateSoftwareTokenClient.mockReturnValueOnce(
+			mockAssociateSoftwareToken,
+		);
 	});
 
 	afterEach(() => {
 		mockAssociateSoftwareToken.mockReset();
 		mockFetchAuthSession.mockClear();
+		mockCreateAssociateSoftwareTokenClient.mockClear();
 	});
 
 	it('setUpTOTP API should call the UserPoolClient and should return a TOTPSetupDetails', async () => {
@@ -62,6 +74,26 @@ describe('setUpTOTP', () => {
 		);
 		expect(result.sharedSecret).toEqual(secretCode);
 		expect(result.getSetupUri('appName', 'amplify')).toBeInstanceOf(URL);
+	});
+
+	it('invokes mockCreateCognitoUserPoolEndpointResolver with expected endpointOverride', async () => {
+		const expectedUserPoolEndpoint = 'https://my-custom-endpoint.com';
+		jest.mocked(Amplify.getConfig).mockReturnValueOnce({
+			Auth: {
+				Cognito: {
+					userPoolClientId: '111111-aaaaa-42d8-891d-ee81a1549398',
+					userPoolId: 'us-west-2_zzzzz',
+					identityPoolId: 'us-west-2:xxxxxx',
+					userPoolEndpoint: expectedUserPoolEndpoint,
+				},
+			},
+		});
+
+		await setUpTOTP();
+
+		expect(mockCreateCognitoUserPoolEndpointResolver).toHaveBeenCalledWith({
+			endpointOverride: expectedUserPoolEndpoint,
+		});
 	});
 
 	it('should throw an error when service returns an error response', async () => {