+++ packages/auth/__tests__/providers/cognito/signInWithUserPassword.test.ts
@@ -6,12 +6,12 @@ import { Amplify } from 'aws-amplify';
 import { signIn } from '../../../src/providers/cognito';
 import * as initiateAuthHelpers from '../../../src/providers/cognito/utils/signInHelpers';
 import { signInWithUserPassword } from '../../../src/providers/cognito/apis/signInWithUserPassword';
-import { RespondToAuthChallengeCommandOutput } from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider/types';
 import {
 	cognitoUserPoolsTokenProvider,
 	tokenOrchestrator,
 } from '../../../src/providers/cognito/tokenProvider';
-import * as clients from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider';
+import { createInitiateAuthClient } from '../../../src/foundation/factories/serviceClients/cognitoIdentityProvider';
+import { RespondToAuthChallengeCommandOutput } from '../../../src/foundation/factories/serviceClients/cognitoIdentityProvider/types';
 
 import { authAPITestParams } from './testUtils/authApiTestParams';
 
@@ -20,20 +20,27 @@ jest.mock('@aws-amplify/core/internals/utils', () => ({
 	...jest.requireActual('@aws-amplify/core/internals/utils'),
 	isBrowser: jest.fn(() => false),
 }));
+jest.mock(
+	'../../../src/foundation/factories/serviceClients/cognitoIdentityProvider',
+);
 
 const authConfig = {
 	Cognito: {
 		userPoolClientId: '111111-aaaaa-42d8-891d-ee81a1549398',
 		userPoolId: 'us-west-2_zzzzz',
+		userPoolEndpoint: 'https://custom-endpoint.com',
 	},
 };
 
-cognitoUserPoolsTokenProvider.setAuthConfig(authConfig);
-Amplify.configure({
-	Auth: authConfig,
-});
 describe('signIn API happy path cases', () => {
 	let handleUserPasswordFlowSpy: jest.SpyInstance;
+
+	beforeAll(() => {
+		Amplify.configure({
+			Auth: authConfig,
+		});
+		cognitoUserPoolsTokenProvider.setAuthConfig(authConfig);
+	});
 
 	beforeEach(() => {
 		handleUserPasswordFlowSpy = jest
@@ -60,7 +67,7 @@ describe('signIn API happy path cases', () => {
 		expect(handleUserPasswordFlowSpy).toHaveBeenCalledTimes(1);
 	});
 
-	test('handleUserPasswordAuthFlow should be called with clientMetada from request', async () => {
+	test('handleUserPasswordAuthFlow should be called with clientMetadata from request', async () => {
 		const { username } = authAPITestParams.user1;
 		const { password } = authAPITestParams.user1;
 		await signInWithUserPassword({
@@ -79,22 +86,19 @@ describe('signIn API happy path cases', () => {
 });
 
 describe('Cognito ASF', () => {
-	let initiateAuthSpy: jest.SpyInstance;
+	const mockInitiateAuth = jest.fn();
+	const mockCreateInitiateAuthClient = jest.mocked(createInitiateAuthClient);
 
-	afterAll(() => {
-		jest.restoreAllMocks();
-	});
 	beforeEach(() => {
-		initiateAuthSpy = jest
-			.spyOn(clients, 'initiateAuth')
-			.mockImplementationOnce(async () => ({
-				ChallengeName: 'SRP_AUTH',
-				Session: '1234234232',
-				$metadata: {},
-				ChallengeParameters: {
-					USER_ID_FOR_SRP: authAPITestParams.user1.username,
-				},
-			}));
+		mockInitiateAuth.mockResolvedValueOnce({
+			ChallengeName: 'SRP_AUTH',
+			Session: '1234234232',
+			$metadata: {},
+			ChallengeParameters: {
+				USER_ID_FOR_SRP: authAPITestParams.user1.username,
+			},
+		});
+		mockCreateInitiateAuthClient.mockReturnValueOnce(mockInitiateAuth);
 		// load Cognito ASF polyfill
 		(window as any).AmazonCognitoAdvancedSecurityData = {
 			getData() {
@@ -104,7 +108,8 @@ describe('Cognito ASF', () => {
 	});
 
 	afterEach(() => {
-		initiateAuthSpy.mockClear();
+		mockInitiateAuth.mockClear();
+		mockCreateInitiateAuthClient.mockClear();
 		(window as any).AmazonCognitoAdvancedSecurityData = undefined;
 	});
 
@@ -120,7 +125,7 @@ describe('Cognito ASF', () => {
 		} catch (_) {
 			// only want to test the contents
 		}
-		expect(initiateAuthSpy).toHaveBeenCalledWith(
+		expect(mockInitiateAuth).toHaveBeenCalledWith(
 			expect.objectContaining({
 				region: 'us-west-2',
 			}),