+++ packages/auth/__tests__/providers/cognito/updateMFAPreference.test.ts
@@ -8,10 +8,11 @@ import {
 	UpdateMFAPreferenceInput,
 	updateMFAPreference,
 } from '../../../src/providers/cognito';
-import { setUserMFAPreference } from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider';
 import { AuthError } from '../../../src/errors/AuthError';
 import { SetUserMFAPreferenceException } from '../../../src/providers/cognito/types/errors';
 import { getMFASettings } from '../../../src/providers/cognito/apis/updateMFAPreference';
+import { createSetUserMFAPreferenceClient } from '../../../src/foundation/factories/serviceClients/cognitoIdentityProvider';
+import { createCognitoUserPoolEndpointResolver } from '../../../src/providers/cognito/factories';
 
 import { getMockError, mockAccessToken } from './testUtils/data';
 import { setUpGetConfig } from './testUtils/setUpGetConfig';
@@ -25,8 +26,9 @@ jest.mock('@aws-amplify/core/internals/utils', () => ({
 	isBrowser: jest.fn(() => false),
 }));
 jest.mock(
-	'../../../src/providers/cognito/utils/clients/CognitoIdentityProvider',
+	'../../../src/foundation/factories/serviceClients/cognitoIdentityProvider',
 );
+jest.mock('../../../src/providers/cognito/factories');
 
 const mfaChoices: UpdateMFAPreferenceInput[] = [
 	{ sms: 'DISABLED', totp: 'DISABLED' },
@@ -51,7 +53,13 @@ const mfaChoices: UpdateMFAPreferenceInput[] = [
 describe('updateMFAPreference', () => {
 	// assert mocks
 	const mockFetchAuthSession = fetchAuthSession as jest.Mock;
-	const mockSetUserMFAPreference = setUserMFAPreference as jest.Mock;
+	const mockSetUserMFAPreference = jest.fn();
+	const mockCreateSetUserMFAPreferenceClient = jest.mocked(
+		createSetUserMFAPreferenceClient,
+	);
+	const mockCreateCognitoUserPoolEndpointResolver = jest.mocked(
+		createCognitoUserPoolEndpointResolver,
+	);
 
 	beforeAll(() => {
 		setUpGetConfig(Amplify);
@@ -62,11 +70,15 @@ describe('updateMFAPreference', () => {
 
 	beforeEach(() => {
 		mockSetUserMFAPreference.mockResolvedValue({});
+		mockCreateSetUserMFAPreferenceClient.mockReturnValueOnce(
+			mockSetUserMFAPreference,
+		);
 	});
 
 	afterEach(() => {
 		mockSetUserMFAPreference.mockReset();
 		mockFetchAuthSession.mockClear();
+		mockCreateSetUserMFAPreferenceClient.mockClear();
 	});
 
 	it.each(mfaChoices)(
@@ -87,6 +99,25 @@ describe('updateMFAPreference', () => {
 			);
 		},
 	);
+
+	it('invokes mockCreateCognitoUserPoolEndpointResolver with expected endpointOverride', async () => {
+		const expectedUserPoolEndpoint = 'https://my-custom-endpoint.com';
+		jest.mocked(Amplify.getConfig).mockReturnValueOnce({
+			Auth: {
+				Cognito: {
+					userPoolClientId: '111111-aaaaa-42d8-891d-ee81a1549398',
+					userPoolId: 'us-west-2_zzzzz',
+					identityPoolId: 'us-west-2:xxxxxx',
+					userPoolEndpoint: expectedUserPoolEndpoint,
+				},
+			},
+		});
+		await updateMFAPreference(mfaChoices[0]);
+
+		expect(mockCreateCognitoUserPoolEndpointResolver).toHaveBeenCalledWith({
+			endpointOverride: expectedUserPoolEndpoint,
+		});
+	});
 
 	it('should throw an error when service returns an error response', async () => {
 		expect.assertions(2);