+++ packages/auth/src/providers/cognito/apis/setUpTOTP.ts
@@ -14,10 +14,11 @@ import {
 } from '../types/errors';
 import { SetUpTOTPOutput } from '../types';
 import { getTOTPSetupDetails } from '../utils/signInHelpers';
-import { associateSoftwareToken } from '../utils/clients/CognitoIdentityProvider';
-import { getRegion } from '../utils/clients/CognitoIdentityProvider/utils';
+import { getRegionFromUserPoolId } from '../../../foundation/parsers';
 import { assertAuthTokens } from '../utils/types';
 import { getAuthUserAgentValue } from '../../../utils';
+import { createAssociateSoftwareTokenClient } from '../../../foundation/factories/serviceClients/cognitoIdentityProvider';
+import { createCognitoUserPoolEndpointResolver } from '../factories';
 
 /**
  * Sets up TOTP for the user.
@@ -30,12 +31,18 @@ import { getAuthUserAgentValue } from '../../../utils';
 export async function setUpTOTP(): Promise<SetUpTOTPOutput> {
 	const authConfig = Amplify.getConfig().Auth?.Cognito;
 	assertTokenProviderConfig(authConfig);
+	const { userPoolEndpoint, userPoolId } = authConfig;
 	const { tokens } = await fetchAuthSession({ forceRefresh: false });
 	assertAuthTokens(tokens);
 	const username = tokens.idToken?.payload['cognito:username'] ?? '';
+	const associateSoftwareToken = createAssociateSoftwareTokenClient({
+		endpointResolver: createCognitoUserPoolEndpointResolver({
+			endpointOverride: userPoolEndpoint,
+		}),
+	});
 	const { SecretCode } = await associateSoftwareToken(
 		{
-			region: getRegion(authConfig.userPoolId),
+			region: getRegionFromUserPoolId(userPoolId),
 			userAgentValue: getAuthUserAgentValue(AuthAction.SetUpTOTP),
 		},
 		{