+++ packages/storage/src/providers/s3/types/options.ts
@@ -8,14 +8,31 @@ import { TransferProgressEvent } from '../../../types';
 import {
 	StorageListAllOptions,
 	StorageListPaginateOptions,
+	StorageSubpathStrategy,
 } from '../../../types/options';
 
+export interface BucketInfo {
+	bucketName: string;
+	region: string;
+}
+
+export type StorageBucket = string | BucketInfo;
 interface CommonOptions {
 	/**
 	 * Whether to use accelerate endpoint.
 	 * @default false
 	 */
 	useAccelerateEndpoint?: boolean;
+	bucket?: StorageBucket;
+}
+
+/**
+ * Represents the content disposition of a file.
+ * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Disposition
+ */
+export interface ContentDisposition {
+	type: 'attachment' | 'inline';
+	filename?: string;
 }
 
 /** @deprecated This may be removed in the next major version. */
@@ -70,19 +87,17 @@ export type RemoveOptions = WriteOptions & CommonOptions;
  * @deprecated Use {@link ListAllWithPathOptions} instead.
  * Input options type with prefix for S3 list all API.
  */
-export type ListAllWithPrefixOptions = Omit<
-	StorageListAllOptions & ReadOptions & CommonOptions,
-	'subpathStrategy'
->;
+export type ListAllWithPrefixOptions = StorageListAllOptions &
+	ReadOptions &
+	CommonOptions;
 
 /**
  * @deprecated Use {@link ListPaginateWithPathOptions} instead.
  * Input options type with prefix for S3 list API to paginate items.
  */
-export type ListPaginateWithPrefixOptions = Omit<
-	StorageListPaginateOptions & ReadOptions & CommonOptions,
-	'subpathStrategy'
->;
+export type ListPaginateWithPrefixOptions = StorageListPaginateOptions &
+	ReadOptions &
+	CommonOptions;
 
 /**
  * Input options type with path for S3 list all API.
@@ -91,7 +106,9 @@ export type ListAllWithPathOptions = Omit<
 	StorageListAllOptions,
 	'accessLevel'
 > &
-	CommonOptions;
+	CommonOptions & {
+		subpathStrategy?: StorageSubpathStrategy;
+	};
 
 /**
  * Input options type with path for S3 list API to paginate items.
@@ -100,7 +117,9 @@ export type ListPaginateWithPathOptions = Omit<
 	StorageListPaginateOptions,
 	'accessLevel'
 > &
-	CommonOptions;
+	CommonOptions & {
+		subpathStrategy?: StorageSubpathStrategy;
+	};
 
 /**
  * Input options type for S3 getUrl API.
@@ -116,6 +135,19 @@ export type GetUrlOptions = CommonOptions & {
 	 * @default 900 (15 minutes)
 	 */
 	expiresIn?: number;
+	/**
+	 * The default content-disposition header value of the file when downloading it.
+	 *   If a string is provided, it will be used as-is.
+	 *   If an object is provided, it will be used to construct the header value
+	 *   based on the ContentDisposition type definition.
+	 * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Disposition
+	 */
+	contentDisposition?: ContentDisposition | string;
+	/**
+	 * The content-type header value of the file when downloading it.
+	 * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type
+	 */
+	contentType?: string;
 };
 
 /** @deprecated Use {@link GetUrlWithPathOptions} instead. */
@@ -137,9 +169,12 @@ export type UploadDataOptions = CommonOptions &
 	TransferOptions & {
 		/**
 		 * The default content-disposition header value of the file when downloading it.
+		 *   If a string is provided, it will be used as-is.
+		 *   If an object is provided, it will be used to construct the header value
+		 *   based on the ContentDisposition type definition.
 		 * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Disposition
 		 */
-		contentDisposition?: string;
+		contentDisposition?: ContentDisposition | string;
 		/**
 		 * The default content-encoding header value of the file when downloading it.
 		 * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Encoding
@@ -165,13 +200,22 @@ export type UploadDataWithPathOptions = UploadDataOptions;
 export type CopySourceWithKeyOptions = ReadOptions & {
 	/** @deprecated This may be removed in the next major version. */
 	key: string;
+	bucket?: StorageBucket;
 };
 
 /** @deprecated This may be removed in the next major version. */
 export type CopyDestinationWithKeyOptions = WriteOptions & {
 	/** @deprecated This may be removed in the next major version. */
 	key: string;
+	bucket?: StorageBucket;
 };
+
+export interface CopyWithPathSourceOptions {
+	bucket?: StorageBucket;
+}
+export interface CopyWithPathDestinationOptions {
+	bucket?: StorageBucket;
+}
 
 /**
  * Internal only type for S3 API handlers' config parameter.