+++ packages/adapter-nextjs/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.17](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.16...@aws-amplify/adapter-nextjs@1.2.17) (2024-09-04)
+
+**Note:** Version bump only for package @aws-amplify/adapter-nextjs
+
+## [1.2.16](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.15...@aws-amplify/adapter-nextjs@1.2.16) (2024-09-03)
+
+### Bug Fixes
+
+- **adapter-nextjs:** duplicate response Set-Cookie headers in pages router ([#13765](https://github.com/aws-amplify/amplify-js/issues/13765)) ([3fedf63](https://github.com/aws-amplify/amplify-js/commit/3fedf6347823611ef5e28554911cf65c1419fce5))
+
+## [1.2.15](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.14...@aws-amplify/adapter-nextjs@1.2.15) (2024-08-26)
+
+**Note:** Version bump only for package @aws-amplify/adapter-nextjs
+
+## [1.2.14](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.13...@aws-amplify/adapter-nextjs@1.2.14) (2024-08-21)
+
+**Note:** Version bump only for package @aws-amplify/adapter-nextjs
+
+## [1.2.13](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.12...@aws-amplify/adapter-nextjs@1.2.13) (2024-08-15)
+
+**Note:** Version bump only for package @aws-amplify/adapter-nextjs
+
+## [1.2.12](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.11...@aws-amplify/adapter-nextjs@1.2.12) (2024-08-07)
+
+**Note:** Version bump only for package @aws-amplify/adapter-nextjs
+
+## [1.2.11](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.10...@aws-amplify/adapter-nextjs@1.2.11) (2024-08-05)
+
+**Note:** Version bump only for package @aws-amplify/adapter-nextjs
+
+## [1.2.10](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.9...@aws-amplify/adapter-nextjs@1.2.10) (2024-07-23)
+
+**Note:** Version bump only for package @aws-amplify/adapter-nextjs
+
+## [1.2.9](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.8...@aws-amplify/adapter-nextjs@1.2.9) (2024-07-22)
+
+### Bug Fixes
+
+- **aws-amplify:** createKeyValueStorageFromCookieStorageAdapter misses default path and secure values ([#13508](https://github.com/aws-amplify/amplify-js/issues/13508)) ([60a559f](https://github.com/aws-amplify/amplify-js/commit/60a559f612e092c76e1499f93547248312db46f2))
+
+## [1.2.8](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.7...@aws-amplify/adapter-nextjs@1.2.8) (2024-07-19)
+
+**Note:** Version bump only for package @aws-amplify/adapter-nextjs
+
 ## [1.2.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.2.6...@aws-amplify/adapter-nextjs@1.2.7) (2024-07-08)
 
 **Note:** Version bump only for package @aws-amplify/adapter-nextjs