+++ packages/api-rest/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.47](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.46...@aws-amplify/api-rest@4.0.47) (2024-09-04)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.46](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.45...@aws-amplify/api-rest@4.0.46) (2024-09-03)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.45](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.44...@aws-amplify/api-rest@4.0.45) (2024-08-26)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.44](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.43...@aws-amplify/api-rest@4.0.44) (2024-08-21)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.43](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.42...@aws-amplify/api-rest@4.0.43) (2024-08-15)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.42](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.41...@aws-amplify/api-rest@4.0.42) (2024-08-07)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.41](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.40...@aws-amplify/api-rest@4.0.41) (2024-08-05)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.40](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.39...@aws-amplify/api-rest@4.0.40) (2024-07-23)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.39](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.38...@aws-amplify/api-rest@4.0.39) (2024-07-22)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.38](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.37...@aws-amplify/api-rest@4.0.38) (2024-07-19)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
 ## [4.0.37](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.36...@aws-amplify/api-rest@4.0.37) (2024-07-08)
 
 **Note:** Version bump only for package @aws-amplify/api-rest